+++ Source/Gapotchenko.GnuTK/Toolkits/Cygwin/CygwinToolkit.cs
@@ -104,13 +104,13 @@ sealed class CygwinToolkit(
             // Unset it if not instructed by a user.
             commandBuilder.Append("unset POSIXLY_CORRECT;");
         }
-        commandBuilder.Append(command);
+        commandBuilder.Append(CygwinIdiosyncrasies.AdjustArgument(command));
         shellArguments.Add(commandBuilder.ToString());
 
         if (commandArguments is [])
             shellArguments.Add(shellPath);
         else
-            shellArguments.AddRange(commandArguments);
+            shellArguments.AddRange(commandArguments.Select(CygwinIdiosyncrasies.AdjustArgument));
 
         return ProcessHelper.Execute(psi);
     }