+++ Source/Tests/Shell/subtests/d/01/test.sh
@@ -1,12 +1,17 @@
 #!/bin/sh
-
 set -eu
 
+# ===========================================================================
 echo "Test D01"
-
 # Tests handling of '\' character for command-line arguments.
+# ===========================================================================
 
 export VAR=ABC
+
+# ---------------------------------------------------------------------------
+# Case 1
+# Handling of an escaped '$' symbol.
+# ---------------------------------------------------------------------------
 
 # Case 1.A
 
@@ -36,6 +41,20 @@ if [ "$actual" != "$expected" ]; then
     exit 2
 fi
 
+# Case 1.D
+
+actual=$(gnu-tk.sh -l echo '$VAR')
+
+if [ "$actual" != "$expected" ]; then
+    echo "Unexpected 1.D: $actual"
+    exit 2
+fi
+
+# ---------------------------------------------------------------------------
+# Case 2
+# Handling of an escaped '\' symbol.
+# ---------------------------------------------------------------------------
+
 # Case 2.A
 
 actual=$(sh -c 'echo "\\$VAR"')
@@ -43,5 +62,23 @@ expected='\ABC'
 
 if [ "$actual" != "$expected" ]; then
     echo "Unexpected 2.A: $actual"
+    exit 2
+fi
+
+# Case 2.B
+
+actual=$(gnu-tk.sh -c 'VAR=ABC; echo "\\$VAR"')
+
+if [ "$actual" != "$expected" ]; then
+    echo "Unexpected 2.B: $actual"
+    exit 2
+fi
+
+# Case 2.C
+
+actual=$(gnu-tk.sh -l sh -c 'VAR=ABC; echo "\\$VAR"')
+
+if [ "$actual" != "$expected" ]; then
+    echo "Unexpected 2.C: $actual"
     exit 2
 fi